visionly.ai/requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pathlib>=1.0.1
colorama>=0.4.6

# API-specific dependencies
pytrends==4.9.2
yfinance>=0.2.36
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
apify-client>=1.6.0
meteostat>=1.6.5
nltk>=3.8.1
wordcloud>=1.9.3
pmdarima>=2.0.4
supabase>=2.3.5
scipy>=1.11.0

# Data processing
python-dateutil>=2.8.2
pytz>=2023.3

# Development tools
black>=23.7.0
pylint>=2.17.5
pytest>=7.4.0
